dom/u2f/tests/test_appid_facet.html

<!DOCTYPE html>
<meta charset=utf-8>
<head>
  <title>FIDO U2F: AppID / FacetID behavior</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="u2futil.js"></script>
  <script type="text/javascript" src="pkijs/common.js"></script>
  <script type="text/javascript" src="pkijs/asn1.js"></script>
  <script type="text/javascript" src="pkijs/x509_schema.js"></script>
  <script type="text/javascript" src="pkijs/x509_simpl.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<h1>FIDO U2F: AppID / FacetID behavior</h1>
<ul>
  <li><a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1231681">Mozilla Bug 1231681 (initial implementation)</a></li>
  <li><a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1245527">Mozilla Bug 1245527 (hardware rewrite)</a></li>
</ul>

<pre id="log"></pre>

<script class="testbody" type="text/javascript">
"use strict";

var version = "U2F_V2";
var challenge = new Uint8Array(16);

add_task(async function(){
  await promiseU2FRegister(null, [{
      version: version,
      challenge: bytesToBase64UrlSafe(challenge),
    }], [], function(res){
      is(res.errorCode, 0, "Null AppID should work.");
  });

  await promiseU2FRegister("", [{
      version: version,
      challenge: bytesToBase64UrlSafe(challenge),
    }], [], function(res){
      is(res.errorCode, 0, "Empty AppID should work.");
  });

  await promiseU2FRegister("http://example.com/appId", [{
      version: version,
      challenge: bytesToBase64UrlSafe(challenge),
    }], [], function(res){
      isnot(res.errorCode, 0, "HTTP scheme is disallowed");
  });

  await promiseU2FRegister("https://example.com/appId", [{
      version: version,
      challenge: bytesToBase64UrlSafe(challenge),
    }], [], function(res){
      is(res.errorCode, 0, "HTTPS origin for example.com should work");
  });

  await promiseU2FRegister(window.location.origin + "/otherAppId", [{
      version: version,
      challenge: bytesToBase64UrlSafe(challenge),
    }], [], function(res){
      is(res.errorCode, 0, "Direct window origin should work");
  });
})

</script>
</body>
</html>
